Refactoring the code and moving shared types and functions into the lib folder
