+++ game/src/routes/[[num]]/+page.ts
@@ -1,5 +1,6 @@
+import { isGameOver } from '$lib/api';
+import { Game } from '../game';
 import type { PageLoad } from './$types';
-import { Game, NUMBER_TRIES } from '../game';
 
 export const load = (({ params }) => {
 	const numberOfGames = Math.max(1, Math.min(+(params.num ?? 1), 4));
@@ -20,15 +21,15 @@ export const load = (({ params }) => {
 		guesses: game.guesses,
 
 		/**
-		 * An array of strings like '__x_c' corresponding to the guesses, where 'x' means
-		 * an exact match, and 'c' means a close match (right letter, wrong place)
+		 * Indexed object mapping a game to an array of strings like '__x_c' corresponding to the guesses,
+		 * where 'x' means an exact match, and 'c' means a close match (right letter, wrong place)
 		 */
 		hints: game.hints,
 
 		/**
 		 * The correct answer, revealed if the game is over
 		 */
-		answers: game.hints[0].findLastIndex((guess) => !!guess) >= game.numberOfGames + NUMBER_TRIES - 1
+		answers: isGameOver(game.hints, game.numberOfGames)
 			? game.answers
 			: Object.keys(game.wordIndices).reduce<{ [gameIndex: string]: null }>((acc, currentIndex) => {
 				acc[currentIndex] = null
